Increase error message display height
Prior to this commit, long error messages would only
display on a window of one line, forcing the user to scroll
through, viewing only a line at a time. This commit increases
the height of the box to a max of 10 lines, so that more
of the message can be viewed at once.